requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.3
anthropic==0.5.0
gunicorn==21.2.0
pytest==7.4.2
pydantic==2.4.2
celery==5.3.4
redis==5.0.0
numpy==1.24.3
pandas==2.0.3 